Extract advanced settings section
